Split overall GUI and file editor
