+++ DarkScript3/StyleConfig.cs
@@ -1,11 +1,5 @@
 using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
 namespace DarkScript3
@@ -14,28 +8,23 @@ namespace DarkScript3
     {
         public Font FontSetting { get; set; }
 
-        public StyleConfig(GUI ds3, Font font)
+        public StyleConfig()
         {
             InitializeComponent();
             StartPosition = FormStartPosition.CenterParent;
-            commentSetting.Init("Comment", GUI.TextStyles.Comment.ForeBrush);
-            stringSetting.Init("String", GUI.TextStyles.String.ForeBrush);
-            keywordSetting.Init("Keyword", GUI.TextStyles.Keyword.ForeBrush);
-            ttKeywordSetting.Init("Keyword (tooltips)", GUI.TextStyles.ToolTipKeyword.ForeBrush);
-            enumPropSetting.Init("Enum Property", GUI.TextStyles.EnumProperty.ForeBrush);
-            globalConstSetting.Init("Global Enum Constant", GUI.TextStyles.EnumConstant.ForeBrush);
-            numberSetting.Init("Number", GUI.TextStyles.Number.ForeBrush);
-            toolTipEnumType.Init("Enum Type", GUI.TextStyles.EnumType.ForeBrush);
-            backgroundSetting.Init("Background", ds3.editor.BackColor);
-            highlightSetting.Init("Highlight", ds3.editor.SelectionColor);
-            highlightBoxSetting.Init("Highlight Box", GUI.TextStyles.HighlightToken.BorderPen.Brush);
-            plainSetting.Init("Default Text", ds3.editor.ForeColor);
-            selectFont.Font = FontSetting = (Font)font.Clone();
-        }
-
-        private void StyleConfig_Load(object sender, EventArgs e)
-        {
-            
+            commentSetting.Init("Comment", TextStyles.Comment.ForeBrush);
+            stringSetting.Init("String", TextStyles.String.ForeBrush);
+            keywordSetting.Init("Keyword", TextStyles.Keyword.ForeBrush);
+            ttKeywordSetting.Init("Keyword (tooltips)", TextStyles.ToolTipKeyword.ForeBrush);
+            enumPropSetting.Init("Enum Property", TextStyles.EnumProperty.ForeBrush);
+            globalConstSetting.Init("Global Enum Constant", TextStyles.EnumConstant.ForeBrush);
+            numberSetting.Init("Number", TextStyles.Number.ForeBrush);
+            toolTipEnumType.Init("Enum Type", TextStyles.EnumType.ForeBrush);
+            backgroundSetting.Init("Background", TextStyles.BackColor);
+            highlightSetting.Init("Highlight", TextStyles.SelectionColor);
+            highlightBoxSetting.Init("Highlight Box", TextStyles.HighlightToken.BorderPen.Brush);
+            plainSetting.Init("Default Text", TextStyles.ForeColor);
+            selectFont.Font = FontSetting = (Font)TextStyles.Font.Clone();
         }
 
         private void okBtn_Click(object sender, EventArgs e)
